+++ tests/api/health_check.rs
@@ -1,0 +1,19 @@
+use crate::helpers::spawn_app;
+
+#[tokio::test]
+async fn health_check_works() {
+    // Arrange
+    let app = spawn_app().await;
+    let client = reqwest::Client::new();
+
+    // Act
+    let response = client
+        .get(&format!("{}/health_check", &app.address))
+        .send()
+        .await
+        .expect("Failed to execute");
+
+    // Assert
+    assert!(response.status().is_success());
+    assert_eq!(Some(0), response.content_length());
+}
